Restructured and simplified creation CashMachine, fixed CashMachineTest errors
