+++ src/test/java/com/pashovpa/CashMachineTest.java
@@ -1,7 +1,9 @@
 package com.pashovpa;
 
+import com.pashovpa.firstproblem.CashMachine;
 import org.junit.jupiter.api.Test;
 
+import java.io.ByteArrayInputStream;
 import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,8 +12,11 @@ public class CashMachineTest {
 
   @Test
   public void exchange() {
-    System.out.println("Such a test is expected: amount = 5, values = 3 2");
-    var expected = new CashMachine().exchange();
+    CashMachine test = new CashMachine();
+
+    ByteArrayInputStream in = new ByteArrayInputStream("5\n 3 2\n".getBytes());
+    test.inputValues(in);
+    var expected = test.exchange();
 
     ArrayList<Long> actualSubList = new ArrayList<>();
     actualSubList.add(3L);
@@ -22,4 +27,3 @@ public class CashMachineTest {
     assertEquals(expected, actual);
   }
 }
-
